requirements.txt
amqp==5.2.0
anyjson==0.3.3
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.4
Django==5.0.3
django-celery==3.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.5
packaging==24.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
